refactor : mypage 내 정보 가지오기 emil값 추가
